requirements.txt
absl-py==0.8.1
alembic==1.3.1
antlr4-python3-runtime==4.8
appdirs==1.4.3
astor==0.8.0
astroid==2.3.3
attrs==19.3.0
backcall==0.1.0
black==19.10b0
bleach==3.1.0
bokeh==1.4.0
click==7.1.1
cliff==2.16.0
cloudpickle==1.3.0
cmaes==0.6.0
cmd2==0.8.9
colorlog==4.0.2
coolname==1.1.0
cycler==0.10.0
dataclasses==0.7
decorator==4.4.1
defusedxml==0.6.0
entrypoints==0.3
gast==0.2.2
google-pasta==0.1.8
graphviz==0.13.2
grpcio==1.25.0
h5py==2.10.0
hydra-core==1.0.0rc2
imageio==2.6.1
importlib-metadata==1.3.0
importlib-resources==3.0.0
ipykernel==5.1.3
ipython==7.10.2
ipython-genutils==0.2.0
ipywidgets==7.5.1
isort==4.3.21
jedi==0.15.1
Jinja2==2.10.3
joblib==0.14.1
json5==0.8.5
jsonschema==3.2.0
jupyter==1.0.0
jupyter-client==5.3.4
jupyter-console==6.1.0
jupyter-core==4.6.1
jupyterlab==1.2.4
jupyterlab-server==1.0.6
Keras==2.3.1
Keras-Applications==1.0.8
Keras-Preprocessing==1.1.0
kiwisolver==1.1.0
lazy-object-proxy==1.4.3
llvmlite==0.30.0
Mako==1.1.0
Markdown==3.1.1
MarkupSafe==1.1.1
matplotlib==3.1.2
mccabe==0.6.1
mistune==0.8.4
more-itertools==8.0.2
mypy==0.761
mypy-extensions==0.4.3
nbconvert==5.6.1
nbformat==4.4.0
networkx==2.4
notebook==6.0.2
numpy==1.17.4
objgraph==3.4.1
omegaconf==2.0.1rc11
opencv-python==4.1.2.30
opt-einsum==3.1.0
optional-django==0.1.0
optuna==2.0.0
packaging==19.2
pandas==1.1.0
pandocfilters==1.4.2
parso==0.5.2
pathspec==0.7.0
pbr==5.4.4
pexpect==4.7.0
pickleshare==0.7.5
Pillow==6.2.1
prettytable==0.7.2
prometheus-client==0.7.1
prompt-toolkit==3.0.2
protobuf==3.11.0
psutil==5.7.0
ptyprocess==0.6.0
Pygments==2.5.2
pylint==2.4.4
pyparsing==2.4.5
pyperclip==1.7.0
pyrsistent==0.15.6
python-dateutil==2.8.1
python-editor==1.0.4
pytz==2019.3
PyWavelets==1.1.1
PyYAML==5.2
pyzmq==18.1.1
qtconsole==4.6.0
regex==2020.2.20
rope==0.16.0
scikit-image==0.16.2
scikit-learn==0.22.2.post1
scipy==1.3.3
seaborn==0.9.0
Send2Trash==1.5.0
six==1.13.0
SQLAlchemy==1.3.11
stevedore==1.31.0
tensorboard==1.15.0
tensorflow-compression==1.3
tensorflow-estimator==1.15.1
tensorflow-gpu==1.15.2
tensorflow-probability==0.7.0
termcolor==1.1.0
terminado==0.8.3
testpath==0.4.4
toml==0.10.0
tornado==6.0.3
tqdm==4.39.0
traitlets==4.3.3
typed-ast==1.4.1
typing==3.7.4.1
typing-extensions==3.7.4.1
wcwidth==0.1.7
webencodings==0.5.1
Werkzeug==0.16.0
widgetsnbextension==3.5.1
wrapt==1.11.2
zipp==0.6.0
